+++ CHANGELOG.md
@@ -11,9 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added change log.
 
 ### Changed
-- Implemented support for strings in `head`, `last` and `tail`.
+- Implemented support for strings in `head`, `last` and `at`.
 - `at`, `head` and `last` now support empty inputs.
 - Improved return types of `opt` (and its variants), `at` and `chainToOpt`.
 
 ### Removed
+- Removed deprecated `optOrCrash` (use `someOrCrash`).
 - Removed deprecated `orElseOpt` (use `alt` instead).